replace leadership settings with controller-backed charm state
Using leadership settings doesn't work very well as units
are coming and going, since the db relation is between units.

What can and does happen is that as unit 0 is going away, it
holds leadership and zaps the settings while processing its
relation departed hooks.  Unit 1 is coming up, and processing
its own relation joined hooks, but since it does not hold
leadership, it cannot store db_uri, etc. anywhere and when
it finally does become the leader, it sees that the settings
are empty and assumes that the DB relation has not been made.