ensure pods are recycled when the licence changes
We do this by adding a crc32 of the licence contents to the name of the secret.
Since this is referred to by the pod spec, changing the secret will change the
pod spec, which will cause new pods to be spawned, which will cause Mattermost
to read and update the licence.